Extract nominal and classifier detector family
